feature/textInput - промежуточный коммит

+++ src/handlers/messageHandlers/adCreation/adCreationSteps/descriptionStep.ts
@@ -1,8 +1,8 @@
-import { Message } from "@telegraf/types";
-import { STEPS_ENUM } from "constants/config";
-import { CHOOSE_MESSAGES, ERROR_MESSAGES } from "constants/messages";
-import { updateAdvertisementDraft } from "services/advertismentDraft";
-import { Context } from "telegraf";
+import { Message } from '@telegraf/types';
+import { STEPS_ENUM } from 'constants/config';
+import { CHOOSE_MESSAGES, ERROR_MESSAGES } from 'constants/messages';
+import { updateAdvertisementDraft } from 'services/advertismentDraft';
+import { Context } from 'telegraf';
 
 export const handleDescriptionStep = async (ctx: Context) => {
   try {
@@ -13,14 +13,16 @@ export const handleDescriptionStep = async (ctx: Context) => {
 
     const text = (ctx.message as Message.TextMessage).text;
 
-    await updateAdvertisementDraft(ctx.from.id.toString(), {
-      currentStep: STEPS_ENUM.PRICE,
-      description: text,
-    });
+    if (text.length <= 255) {
+      await updateAdvertisementDraft(ctx.from.id.toString(), {
+        currentStep: STEPS_ENUM.PRICE,
+        description: text,
+      });
 
-    await ctx.deleteMessage();
-
-    return ctx.reply(CHOOSE_MESSAGES.PRICE);
+      await ctx.deleteMessage();
+      return ctx.reply(CHOOSE_MESSAGES.PRICE);
+    }
+    return ctx.reply(ERROR_MESSAGES.ERROR_WITH_DISCRIPTION);
   } catch (error) {
     return ctx.reply(ERROR_MESSAGES.ERROR);
   }
